견적서 4차수정

+++ src/main/java/com/example/erp/company/dto/CompanyDto.java
@@ -16,7 +16,7 @@ public class CompanyDto {
     private String companyTel;
     private String status;
     private int moneyRecieve;
-    private long money;
+    private int money;
 
     public static CompanyDto toCompanyDto(CompanyEntity companyEntity) {
         CompanyDto companyDto = new CompanyDto();
